feed:实现编辑功能

+++ src/pages/Auth/UserList.tsx
@@ -1,26 +1,48 @@
 import React, { useRef, useState } from 'react';
 import { ActionType, PageContainer, ProColumns, ProTable } from '@ant-design/pro-components';
-import { Button, Space, Switch } from 'antd';
+import { Button, message, Modal, Space, Switch } from 'antd';
 import { PlusOutlined } from '@ant-design/icons';
 import { FormattedMessage, useModel } from '@@/exports';
-import { deleteUser, restoreUser, user } from '@/services/user/api';
+import { addUser, deleteUser, restoreUser, user } from '@/services/user/api';
 import CreateUserModal from '@/pages/Auth/components/CreateUserModal';
+import UpdateForm from '@/pages/TableList/components/UpdateForm';
+
+const handleUpdate = async (fields: API.CurrentUser) => {
+    const hide = message.loading('正在更新');
+    try {
+        await addUser({ ...fields });
+        hide();
+        message.success('添加成功');
+        return true;
+    } catch (error) {
+        hide();
+        message.error('添加失败，请重试!');
+        return false;
+    }
+};
 
 const UserList: React.FC = () => {
     const { initialState } = useModel('@@initialState');
     const [createModalOpen, handleModalOpen] = useState<boolean>(false);
+    const [currentRow, setCurrentRow] = useState<API.CurrentUser>();
     const actionRef = useRef<ActionType>();
     const switchOpen = async (val: boolean, id: number) => {
-        try {
-            if (val) {
-                await restoreUser(id);
-            } else {
-                await deleteUser(id);
-            }
-            actionRef.current?.reload();
-        } catch (e) {
-            actionRef.current?.reload();
-        }
+        Modal.confirm({
+            title: '确认',
+            content: `确认${val ? '启用' : '关闭'}账户吗`,
+            onOk: async () => {
+                try {
+                    if (val) {
+                        await restoreUser(id);
+                    } else {
+                        await deleteUser(id);
+                    }
+                    actionRef.current?.reload();
+                } catch (e) {
+                    actionRef.current?.reload();
+                }
+            },
+        });
     };
     const valueEnum = new Map([
         [0, '是'],
@@ -35,12 +57,23 @@ const UserList: React.FC = () => {
         {
             title: '账号',
             dataIndex: 'username',
+            render: (dom, entity) => {
+                return (
+                    <a
+                        onClick={() => {
+                            setCurrentRow(entity);
+                        }}
+                    >
+                        {dom}
+                    </a>
+                );
+            },
         },
         {
             title: '姓名',
             dataIndex: 'nickname',
             render: (_, record) => {
-                return <span>{record.profile?.nickName}</span>;
+                return <span>{record.profile?.nickname}</span>;
             },
         },
         {
@@ -54,6 +87,12 @@ const UserList: React.FC = () => {
             title: '添加时间',
             valueType: 'dateTime',
             dataIndex: 'createdAt',
+            search: false,
+        },
+        {
+            title: '最后修改时间',
+            valueType: 'dateTime',
+            dataIndex: 'lastModifiedAt',
             search: false,
         },
         {
@@ -78,9 +117,17 @@ const UserList: React.FC = () => {
             render: (_, record) => {
                 return (
                     <Space>
-                        <a href="" onClick={() => {}}>
-                            编辑
-                        </a>
+                        {initialState?.currentUser?.id !== record.id && (
+                            <a
+                                onClick={(event) => {
+                                    event.preventDefault();
+                                    setCurrentRow(record);
+                                    handleModalOpen(true);
+                                }}
+                            >
+                                编辑
+                            </a>
+                        )}
                     </Space>
                 );
             },
@@ -108,6 +155,7 @@ const UserList: React.FC = () => {
                         type="primary"
                         key="primary"
                         onClick={() => {
+                            setCurrentRow(undefined);
                             handleModalOpen(true);
                         }}
                     >
@@ -117,11 +165,14 @@ const UserList: React.FC = () => {
                 request={user}
                 columns={columns}
             />
-            <CreateUserModal
-                createModalOpen={createModalOpen}
-                handleModalOpen={handleModalOpen}
-                actionRef={actionRef}
-            />
+            {createModalOpen && (
+                <CreateUserModal
+                    createModalOpen={createModalOpen}
+                    handleModalOpen={handleModalOpen}
+                    actionRef={actionRef}
+                    currentRow={currentRow}
+                />
+            )}
         </PageContainer>
     );
 };
